Script to Create a New Terraform Task
For development work, I like to have the main outcome of the work as
the 'parent' task, with the subtasks being all the steps involved in
completing the work. There are subtasks I wanted to have that I found
myself creating all the time and that got pretty tedious, so I decided
to automate it with this script.

This script is specifically for working with Terraform, which has a few
extra tasks that are specific to that kind of work.

This script:
* Prompts for and creates the main task with the provided JIRA reference
* Prompts for and creates any subtasks required
* Prompts for the Terraform modules that will be modified as part of
  this work
* Prompts for the new version numbers of the Terraform modules
* Creates merge tasks for the Terraform modules
* Creates tag tasks for the Terraform modules
* Creates tasks for updating the main Terraform repo with new module
  versions
* Creates a merge task for the main Terraform repo

All tasks/subtasks are created using 'work' and 'development' labels.

As part of this, some code from the 'add tasks to section' script was
refactored out into a little common modules.