Create Subtasks for Merge Related Tasks
The merge tasks are extended to have subtasks for creating a pull
request, and in the case of the Terraform task, there are also subtasks
related to running plans for the `prod` and `int` environments.

A lot of duplicate code was moved out into a generic 'lib' module.
The other separate files can just be merged into this module, since it's
just supposed to be a simple little module to avoid duplication. That's
probably preferable to have lots of different little files. These are
just simple scripts, so there isn't much else needed.