New Merge and Admin Subtasks
Today I discovered there were another couple of subtasks that would be
useful to have.

One is to create a task for rebasing master into the branch I'm
attempting to merge (to avoid conflicts on the pull request submission)
and the other is for creating an admin task to close off the JIRA
ticket.