Produce More Template Subtasks
There are several more subtasks created:
* Creating a branch for the main repo
* Creating a branch for any modules used
* Creating post-merge tasks

These are arguably a bit excessive, but it just helps keep me on track
with exactly everything I need to do for the work and there's not much
harm in just having them if they're automated anyway.

More utility functions were adding for branching and merging.

This also includes a ui_get_repos method (commented out at the moment)
that's intended to be used for the more generic development task. I
initially started modifying the UI library for use with this task, but
then discovered I wanted several additions to the Terraform task.