Script to Create a New Development Task
This is similar to the `create-terraform-task` in that the outcome of
the work is defined by the name of the root task and all the subtasks
are the steps necessary for achieving the outcome.

Right now this only supports using a single repository, but it shouldn't
be too hard to extend for using multiple repos.

The script uses basic argument parsing to allow defining the subtasks
in a file. Each line of the file is interpreted as a subtask and you can
add a comment to the task by separating the name and the comment using
the ';' character.