Script for Creating Defect Tasks
This is very similar to the support task, but it allows the selection of
a project.

There's been a change here to filter the project selection based on work
projects rather than including the full list, which is a lot more
convenient to work with for these tasks.

Also made a modification to assign the root task with today's date, as
I'm generally raising these to work on them on the same day.