Script for Shifting Current Day Outline
This script shifts all the tasks in the current day outline, based on
the time it is right now and the duration that's been specified in the
task name. If the current time is before half past the current hour, the
new outline will begin at 30 minutes past, otherwise it will begin on
the next hour. For example, if the time is currently 1315, the new day
outline will begin at 1330, but if the current time is 1335, the new
outline will begin at 1400.

All of the other tasks in the outline will have their start time shifted
based on the duration specified in the previous task. This works because
the tasks in the day outline are always specified as
"Description [X hours]".

It's extremely tedious to do this by hand, especially because the
Todoist GUI adds reminders if you change the time of the task. This
script should be very helpful.

There will need to be modifications to accomodate meetings during a
working day.