Script for Creating a PR Task
This was easier to do in Bash because the Azure CLI has commands on it
that allow retrieval of the details of a PR. I still don't know how to
do that with the Azure Python libraries, as of yet (though really should
learn, as it won't be difficult).

This can be extended later to provide a project ID using a variable, or
maybe even call out to Python to select the project and return the ID.
It would be extremely cumbersome to do that with Bash.