Script for Creating Whitelist Task
Changing people's whitelist entry, or creating new entries, is a very
common task. Since it's a Terraform activity, it could have been done
with the Terraform task script, but whitelisting is always the same
specific action, so it's quicker to create a whitelist task directly.

It will ask for all the whitelist entries to be created or updated
(since fairly often you can do more than one at a time), then it'll
create all the necessary sub tasks.

There were a few things refactored in the task libraries to make the
creation of tasks more generic.

The UI library was also refactored to remove the creation of a task
directly in there. Where JIRA references are used, the task is now
created with a link to the JIRA ticket.

Decided to remove the script descriptions from the README. It's too
cumbersome to update this every time a script is added, and most of them
should be pretty self explanatory, based on the name. If necessary,
documentation should be supplied on the scripts themselves.