Script for Importing a Day Outline
Specifying the day outline can actually be quite tedious to do manually.
This simple script will parse a file where a day outline is specified.
It should be very quick to define the day outline based on another
example. It will be much quicker to compose the day by editing text in
Vim that it will be to use the Todoist GUI.