Allow Specification of Subtasks
This uses a very simple method that only allows for 1 level of subtask,
rather than allowing an arbitrary level using recursion or something
like that. All I need is just a single level.

A task is denoted by a leading '*' and a subtask is denoted with a
leading '-'. The text file can be made more readable using indentation,
but any leading whitespace will be stripped during processing.

Also corrects a bad reference to a function that had changed names.