Use More Generic Branch Reference
Since JIRA is not in use at my new place, I'm changing these tasks so
they'll prompt for a JIRA or branch reference, as opposed to just a
JIRA reference.

The code will also check for the existence of the JIRA_BASE_URL
variable and name the tasks appropriately, and also create (or not)
an admin task to close the JIRA ticket.