Supply Task and Work Type Args
The script for creating a dev task has been extended so you can say
whether the task is for work or is a personal one.

The type of the task can also be specified, e.g. whether it's for Rust
or Python etc. as this will determine whether certain subtasks are
added.