Define some rules for branch names
Sometimes here I forget that it's actually asking for a branch and I supply the name of the task
instead.

Defining a couple of rules for the branch name will help take me out of autopilot and not make the
mistake again.

The branch name won't allow uppercase characters or spaces.