Move the dev module to a lib module
The original reason for naming it `dev` was for that to be one category of automated tasks and that
more would come up over time, but I've never encountered that yet.

Now I'm going to be more generic and just say that these are library functions. Anything else that
gets added can just go there too. I don't think there'll ever be enough for them to be worth
breaking down.