Script to clear the current day
Any personal, work or routine tasks will have their current due date set to null, which will have
the effect of clearing the board for today.

If you were doing this in the early hours of the current day, in an attempt to clear out what are
now yesterday's routine tasks, it might be desirable to reset them to today's date. The `--today`
flag argument can be used for this. This will only apply to the routine tasks. You would need to
plan/reschedule any non-routine work, which is the desirable effect.