Correctly assign due dates
Use a check to only assign the due date if the task is either due today or overdue. Clearing the
date from a recurring task with a future due date had the effect of creating a new date and
pushed the task even further out into the future.

Also added documentation for script usage, with particular emphasis on the `--today` flag.