feat: script for processing a pull request
I'm now going to use conventional commits.

This new script enables the user to process a checklist intended to make sure work is ready for PR,
submits the PR, then updates the dev task with a 'waiting' status and a link to the PR.

The checklist items correspond to subtasks in the merge subtask on a development task. The script
let's the user choose which outstanding dev task they wish to merge.

A couple of new libraries are incorporated for UI/UX. The first is `questionary`, which provides
selection and checklist UIs for the terminal, and the second is 'rich', which provides colours and a
very nice way of indicating there's a long running task.