feat: use questionary for dev task ui
As per the 'submit PR checklist' script, the `questionary` library is now used in the 'create dev
task' script.

The subtasks created as part of a dev task were also updated. There's now a 'PR checklist' task and
separate merge and submit PR subtasks.

The submit PR script needs to be updated for this new structure.