feat: script to save an item to buy
A quick mechanism for saving books or films/tv/documentaries to buy. I initially looked at scraping
the information from the Amazon product item page, but they don't make it easy. For now it was
easier to just enter the information manually, and it also means I can use other sources.

I extended the `create_task` function to make it a bit more generic for adding non-development
items. The use of `black` for formatting also applied consistent formatting to the code.